+++ CHANGES.rst
@@ -8,6 +8,8 @@ Unreleased
 - Added support for `CrateDB bulk operations`_, for improved efficiency on
   DML operations.
 
+- Added deprecation notice about PHP7.
+
 .. _CrateDB bulk operations: https://crate.io/docs/crate/reference/en/latest/interfaces/http.html#bulk-operations
 
 2022/11/29 2.1.4